refactor: add validateFields function in utils folder
